Fix Stateful Metrics in fit_generator with TensorBoard (#10673)
* Fix casting validate generator with stateful metrics.

* Add a stateful metric to Tensboard tests

+++ keras/engine/training_generator.py
@@ -375,7 +375,7 @@ def evaluate_generator(model, generator,
             averages.append(np.average([out[i] for out in outs_per_batch],
                                        weights=batch_sizes))
         else:
-            averages.append(float(outs_per_batch[-1][i]))
+            averages.append(np.float64(outs_per_batch[-1][i]))
     return unpack_singleton(averages)
